feat: Generate extension attributes based on proto annotations
This exposes extension attributes (as static properties) within a static ExtensionAttributes class per directory.
The extension attributes are all expected to be strings, and are not currently used in the formatter - but they can still be accessed afterwards. (Test will come after generation commit.)

Unfortunately, in order to implement this, the generator needed
fairly significant rewriting.